Fixing a numeric / string conversion bug.
Perl's numeric and string data types, being effectively equivalent, are
dumped differently by Data::Dumper.  This itself wouldn't be a problem,
save for the Generator's expectation that it would need to strip off
quoting.  Since this is not consistently the case (and q* quoting would
be broken anyhow), simplifying the Generator to simply use the Dumpered
string as-is turns out to be a win.