Fix spec test suite breakages
The spec test suite (`t/001_spec_compatibility`) was breaking in Perl
versions 5.16 and below on the "Sections" test within `delimiters.yml` and
and the "Nested (Truthy)" test in `sections.yml` with the error

    Can't call method "can" without a package or object reference

The issue was that the `$ctx` variable being used within the `lookup()`
function could be a simple as the value `1`.  The original code would then
attempt to call `->can($field)` on this value, hence causing the error
mentioned above and hence the test failures.  By running the `->can()` call
within an `eval` block, we can catch the error, and if the error occurs,
then specify a false value for the ability to run the `can()` method, which
implements - I believe - the desired behaviour.

However, the ability to run `can()` with or without error is dependent upon
the version of Perl; versions 5.16 and below raise an error, 5.18 and above
return `undef`.  Hence the decision about whether or not the current value
of the `$ctx` variable can run the given `$field` as a method or not needs
to be handled appropriately.  This can be done with named subroutine which
calls the error-handling code for Perl versions 5.16 and below, and for Perl
versions above 5.16 the original `$ctx->can($field)` code can be used.